ToolBox/requirements.txt
telebot==0.0.5
requests==2.27.1
python-dotenv==1.0.1
pillow==11.0.0
PyPDF2==3.0.1
md2tgmd==0.3.9
python-dateutil==2.9.0.post0